backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
websockets==15.0.1
pydantic-settings==2.10.1
python-dotenv==1.1.1
numpy==2.3.2
soundfile==0.13.1
prometheus-client==0.22.1
faster-whisper==1.2.0
silero-vad==5.1.2
google-genai==1.10.0
onnxruntime
kokoro-onnx
httpx
